Fix format string for pack
The format string of struct.pack may not be unicode in earlier python2.7
versions. With 2.7.3 (Debian oldstable) the bug is present. With 2.7.13
(Debian stable) a unicode is accepted (like for python3.4 and 3.5).
Fortunately python3 also accepts a byte-string here, so we use a byte
string for the format for pack. Fixes Github issue #3.